// src/lib/interfaces/maps/ponds.ts
export interface MotorAddData {
  id: number;
  name: string | undefined;
  power: string | undefined;
  location: [number, number];
  motorId: number;
}
export interface PondData {
  id: string;
  coordinates: number[][];
  area: number;
  name: string;
  location: any;
  motors: MotorAddData[];
}
export interface IMapIconsPageProps {
  showResults: boolean;
  pond: PondData | null;
  drawMode: string;
  handleLineStringMode: () => void;
  handlePolygonMode: () => void;
  handleUndo: () => void;
  handleRedo: () => void;
  handleDeletePond: () => void;
  handleEdit: () => void;
  canUndo: boolean | null;
  canRedo: boolean | null;
  isPondEditing: boolean;
}

export interface IAddPondSideBarProps {
  pond: PondData | null;
  setPond: (pond: PondData | null) => void;
  handleAddMotor: () => void;
  isAddingMotor: boolean;
  handleMotorNameChange: (motorId: number, newName: string) => void;
  handleMotorPowerChange: (motorId: number, newPower: string) => void;
  handleDeleteMotor: (motorId: number) => void;
  handleCancel: () => void;
  // handleSavePond: () => void;
  errorMessage: Record<string, string> | null;
  setErrorMessage: Dispatch<SetStateAction<Record<string, string> | null>>;
  // isPending: boolean;
  pondName: string;
  setPondName: Dispatch<SetStateAction<string>>;
  locationCentroid: number[];
  setIsPondAdding: Dispatch<SetStateAction<boolean>>;

  setIsPondEditing: Dispatch<SetStateAction<boolean>>;
  originalPond: any | null;
  isPondEditing: boolean;
  deleteMotor: (payload: { motorId: number; pondId: number }) => void;
  isDeletingMotorPending: boolean;
  addParam: string | null;
  setSelectedLocation?: (location: any) => void;
  setSelectedLocationId?: (id: number | null) => void;
  selectedMotorId: number | null;
  addPondsToMap: () => void;
}
// ...
export interface StarterBox {
  id: number;
  title: string;
  alias_starter_title: string | null;
  status: string
  mac_address: string;
  gateways?: Gateway
}

export interface UserDevices {
  id: number;
  title: string;
  alias_starter_title: string | null;
  connected_nodes: string[];
  capable_motors: number;
}


export interface Faults {
  fault_code: number;
  fault_description: string;
  time_stamp: string;
  power_present: string;
}

export interface StarterBoxParameters {
  id: number;
  starter_id: number;
  motor_id: number;
  current_i1: number;
  current_i2: number;
  current_i3: number;
  line_voltage_vry: number;
  line_voltage_vyb: number;
  line_voltage_vbr: number;
  power_present?: string;
  fault_code?: number;
  fault_description?: string;
}

export interface Motor {
  id: number;
  title: string;
  state: number;
  starter_id: number | null;
  motor_ref_id: string | null;
  status: string
  mode: string
  hp: string;
  motor_coordinates: {
    lat: number;
    lng: number;
  };
  starterBoxParameters: StarterBoxParameters[];
  starterBox: StarterBox;
  faults: Faults;
}